requirements.txt
alabaster==0.7.12
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
idna==2.8
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
packaging==19.0
Pillow==5.4.1
postgres==2.2.2
psycopg2-binary==2.7.7
Pygments==2.3.1
pyparsing==2.3.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-plantuml==0.14
sphinxcontrib-websupport==1.1.0
urllib3==1.24.1
wxPython==4.0.4
